# app-api/4iiz-api/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: 4iiz API
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /climbs:
    x-swagger-router-controller: climbs
    get:
      description: Returns list of Climbs
      operationId: getClimbs
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ClimbsResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      description: Calculates a new Snail climb and records the transaction
      operationId: newClimb
      parameters:
        - name: climb
          in: body
          schema:
            $ref: '#/definitions/ClimbRequest'
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ClimbResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /climbs/{climbId}:
    x-swagger-router-controller: climbs.climb
    get:
      description: Get results of a specific climb
      operationId: getClimb
      produces:
        - application/json
      parameters:
        - name: climbId
          in: path
          description: Get Climb Stats
          type: string
          required: true
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ClimbResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  ClimbsResponse:
    properties:
      id:
        type: string
      timestamp:
        type: string
  ClimbRequest:
    required:
      - h
      - u
      - d
      - f
    properties:
      h:
        type: number
      u:
        type: number
      d:
        type: number
      f:
        type: number
  ClimbResponse:
    properties:
      results:
        type: array
        items:
          $ref: '#/definitions/ClimbDayLog'
  ClimbDayLog:
    properties:
      h:
        type: number
      u:
        type: number
      d:
        type: number
      f:
        type: number
      result:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string